requirements.txt
flask
flask-cors
requests
PyPDF2
faiss-cpu
numpy
sentence-transformers
python-docx
python-pptx
